Bugfix: setEnvVar
If setEnvVar was used twice in a row, the first env var key/value pairs
would be dropped and the second would be kept. This was due to us
pulling `environment` from React's state and the re-render cycle
doesn't occur quick enough for `environment` to have changes from
update 1 of 2.

By pulling directly from localStorage, which is synchronous, it
guarantees the updates took effect