Bugfix: New Routes Show Zero In Response Section
When a workspace is initialized and routes are loaded for the first
time, before requests are made to a route they would show `0` because
of a `status` that defaults to `0`. They would also show `Headers +`
but clicking on the div to expand and see headers would render an empty
box.